refac(flagstats): moved v3 to its own component
